Revert "Add reporting for CI"
This reverts commit c4834f47a1b505833c651e8828d640f3eb4f1aac.

ci_reporter replaces $stdout with a StringIO, which causes our use of #spawn to fail because StringIO isn't a real IO that can be passed to a sub-process.

We would need to pipe all IO if we wanted to allow SippyCup to function in light of a StringIO-replaced $stdout, just like [Poltergeist had to](https://github.com/jonleighton/poltergeist/commit/3365dadfb6242b0b91fe00359ff881e582cc2557), and we do for stderr.