[FEATURE] Split receive_200 into its own method
It was an alias of receive_answer but receive_answer also stores the
record set and the response time duration, which you probably don't
want to do on every 200 received. Also simplify wait_for_answer and
fix the documentation, which stated that these expectations are
optional by default when they are not.