Fix specs under ActiveSupport 4
capture and quietly are deprecated because they are not
thread-safe. That does not matter for these specs and the former is
required for some tests so import the code from ActiveSupport. The
latter isn't strictly required but we might as well bring that across
too.

Since we have a global around block to apply quietly, the individual
invocations of it are not necessary.